requirements.txt
seaborn==0.13.2
pandas==2.3.3
streamlit==1.26.1
loguru==0.7.0
joblib==1.3.2
scikit-learn
matplotlib==3.8.0
numpy==1.26.4
catboost 
xgboost 
lightgbm
pathlib
flask==3.0.3
flask-cors==4.0.0
gunicorn==21.2.0
duckdb
requests==2.31.0
psycopg2-binary